requirements.txt
# SPDX-FileCopyrightText: © 2024 Stacey Adams <stacey.belle.rose@gmail.com>
# SPDX-License-Identifier: Unlicense
psutil >= 5.8.0, < 6.0
python-dotenv==0.21.0
requests >= 2.25.0, < 3.0
rpi_lcd==0.0.3
smbus==1.1.post2
spotipy >= 2.20.0, < 3.0
typing-extensions >= 4.10
